requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt
gunicorn==20.0.4
idna==2.10
psycopg2-binary
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
